other/mark_gen/requirements.txt
croniter==1.4.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
loguru==0.7.2